requirements.txt
dash>=2.16
plotly>=5.22
pandas>=2.0
python-dotenv>=1.0
requests>=2.32
